requirements.txt
google-generativeai
google-api-python-client
pandas
python-dotenv
tqdm
markdown
requests
yt-dlp
